requirements.txt
astroid==2.1.0
confusable-homoglyphs==3.2.0
Django==1.11
django-registration==3.0
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pylint==2.2.2
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
typing==3.6.6
wrapt==1.11.1
